Fix various warning around conversion of types

+++ tutorial/detection/tag/tutorial-apriltag-detector-live-T265-realsense.cpp
@@ -159,7 +159,7 @@ int main(int argc, const char **argv)
   double opt_tag_size = 0.053;
   float opt_tag_quad_decimate = 1.0;
   float opt_tag_decision_margin_threshold = 50;
-  float opt_tag_hamming_distance_threshold = 2;
+  int opt_tag_hamming_distance_threshold = 2;
   int opt_tag_nThreads = 1;
   bool opt_display_tag = false;
   int opt_color_id = -1;
